docs: content change to a single module and export default
